Add rudimentary support for different agent models
